requirements.txt
Flask
pytesseract
opencv-python
numpy
Pillow